+++ core/src/jvmMain/kotlin/zakadabar/stack/backend/data/builtin/role/RoleBackend.kt
@@ -8,6 +8,8 @@ package zakadabar.stack.backend.data.builtin.role
 import io.ktor.routing.*
 import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.transactions.transaction
+import zakadabar.stack.StackRoles
+import zakadabar.stack.backend.authorize
 import zakadabar.stack.backend.data.record.RecordBackend
 import zakadabar.stack.data.builtin.RoleDto
 import zakadabar.stack.util.Executor
@@ -25,12 +27,18 @@ object RoleBackend : RecordBackend<RoleDto>() {
     }
 
     override fun all(executor: Executor) = transaction {
+
+        authorize(executor, StackRoles.securityOfficer)
+
         RoleTable
             .selectAll()
             .map(RoleTable::toDto)
     }
 
     override fun create(executor: Executor, dto: RoleDto) = transaction {
+
+        authorize(executor, StackRoles.securityOfficer)
+
         RoleDao.new {
             name = dto.name
             description = dto.description
@@ -38,10 +46,16 @@ object RoleBackend : RecordBackend<RoleDto>() {
     }
 
     override fun read(executor: Executor, recordId: Long) = transaction {
+
+        authorize(executor, StackRoles.securityOfficer)
+
         RoleDao[recordId].toDto()
     }
 
     override fun update(executor: Executor, dto: RoleDto) = transaction {
+
+        authorize(executor, StackRoles.securityOfficer)
+
         val dao = RoleDao[dto.id]
         with(dao) {
             name = dto.name
@@ -51,6 +65,9 @@ object RoleBackend : RecordBackend<RoleDto>() {
     }
 
     override fun delete(executor: Executor, recordId: Long) {
+
+        authorize(executor, StackRoles.securityOfficer)
+
         RoleDao[recordId].delete()
     }
 }
